+++ Metrics.Tests/Sampling/WeightedSnapshotTests.cs
@@ -1,11 +1,11 @@
-﻿using System;
+using System;
 using System.Linq;
 
 using FluentAssertions;
 
 using Metrics.Sampling;
 
-using Xunit;
+using NUnit.Framework;
 
 namespace Metrics.Tests.Sampling
 {
@@ -23,126 +23,126 @@ namespace Metrics.Tests.Sampling
             return new WeightedSnapshot(values.Length, samples);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_SmallQuantilesAreTheFirstValue()
         {
             snapshot.GetValue(0.0).Should().Be(1.0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_BigQuantilesAreTheLastValue()
         {
             snapshot.GetValue(1.0).Should().Be(5.0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_HasAMedian()
         {
             snapshot.Median.Should().Be(3.0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_HasAp75()
         {
             snapshot.Percentile75.Should().Be(4.0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_HasAp95()
         {
             snapshot.Percentile95.Should().Be(5.0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_HasAp98()
         {
             snapshot.Percentile98.Should().Be(5.0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_HasAp99()
         {
             snapshot.Percentile99.Should().Be(5.0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_HasAp999()
         {
             snapshot.Percentile999.Should().Be(5.0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_HasValues()
         {
             snapshot.Values.Should().Equal(new long[] {1, 2, 3, 4, 5});
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_HasSize()
         {
             snapshot.Size.Should().Be(5);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_CalculatesTheMinimumValue()
         {
             snapshot.Min.Should().Be(1);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_CalculatesTheMaximumValue()
         {
             snapshot.Max.Should().Be(5);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_CalculatesTheMeanValue()
         {
             snapshot.Mean.Should().Be(2.7);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_CalculatesTheStdDev()
         {
             snapshot.StdDev.Should().BeApproximately(1.2688, 0.0001);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_CalculatesAMinOfZeroForAnEmptySnapshot()
         {
             Snapshot snapshot = MakeSanpshot(new long[0], new double[0]);
             snapshot.Min.Should().Be(0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_CalculatesAMaxOfZeroForAnEmptySnapshot()
         {
             Snapshot snapshot = MakeSanpshot(new long[0], new double[0]);
             snapshot.Max.Should().Be(0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_CalculatesAMeanOfZeroForAnEmptySnapshot()
         {
             Snapshot snapshot = MakeSanpshot(new long[0], new double[0]);
             snapshot.Mean.Should().Be(0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_CalculatesAStdDevOfZeroForAnEmptySnapshot()
         {
             Snapshot snapshot = MakeSanpshot(new long[0], new double[0]);
             snapshot.StdDev.Should().Be(0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_CalculatesAStdDevOfZeroForASingletonSnapshot()
         {
             Snapshot snapshot = MakeSanpshot(new long[0], new double[0]);
             snapshot.StdDev.Should().Be(0);
         }
 
-        [Fact]
+        [Test]
         public void WeightedSnapshot_ThrowsOnBadQuantileValue()
         {
             ((Action)(() => snapshot.GetValue(-0.5))).Should().Throw<ArgumentException>();